depreciate lazyPainterResource refactor to asyncPainterResource
