+++ kamel-samples/src/desktopMain/kotlin/io/kamel/samples/SvgSample.kt
@@ -11,7 +11,7 @@ import io.kamel.image.config.Default
 import io.kamel.image.config.LocalKamelConfig
 import io.kamel.image.config.resourcesFetcher
 import io.kamel.image.config.svgDecoder
-import io.kamel.image.lazyPainterResource
+import io.kamel.image.asyncPainterResource
 
 public fun main(): Unit = singleWindowApplication {
     val kamelConfig = KamelConfig {
@@ -21,7 +21,7 @@ public fun main(): Unit = singleWindowApplication {
     }
 
     CompositionLocalProvider(LocalKamelConfig provides kamelConfig) {
-        val painterResource = lazyPainterResource("MR/files/Kotlin.svg")
+        val painterResource = asyncPainterResource("MR/files/Kotlin.svg")
 
         KamelImage(
             painterResource,
